[FEAT] 리사이징 함수 커스텀훅으로 분리 및 테스트 코드 작성 #17
